# docs/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Arc LMS API",
...
    "version": "1.0.0",
    "contact": {
      "name": "Arc LMS Support",
      "email": "support@arc-lms.com",
      "url": "https://arc-lms.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Development server"
    },
    {
      "url": "https://arc-lms.onrender.com/api/v1",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/courses": {
      "get": {
        "tags": ["Courses"],
        "summary": "List courses",
        "operationId": "listCourses",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {"$ref": "#/components/parameters/Page"},
          {"$ref": "#/components/parameters/Limit"},
          {"name": "class_id", "in": "query", "schema": {"type": "string", "format": "uuid"}}
        ],
        "responses": {
          "200": {"description": "Courses retrieved", "content": {"application/json": {"schema": {"type": "object", "properties": {"data": {"type": "array", "items": {"$ref": "#/components/schemas/Course"}}, "pagination": {"$ref": "#/components/schemas/PaginatedResult"}}}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"}
        }
      },
      "post": {
        "tags": ["Courses"],
        "summary": "Create course",
        "operationId": "createCourse",
        "security": [{"BearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateCourseRequest"}}}},
        "responses": {
          "201": {"description": "Course created", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Course"}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"}
        }
      }
    },
    "/courses/{id}": {
      "get": {
        "tags": ["Courses"],
        "summary": "Get course by ID",
        "operationId": "getCourse",
        "security": [{"BearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}],
        "responses": {
          "200": {"description": "Course retrieved", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Course"}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      },
      "put": {
        "tags": ["Courses"],
        "summary": "Update course",
        "operationId": "updateCourse",
        "security": [{"BearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}, {"$ref": "#/components/parameters/IdempotencyKey"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UpdateCourseRequest"}}}},
        "responses": {
          "200": {"description": "Course updated", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Course"}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      },
      "delete": {
        "tags": ["Courses"],
        "summary": "Delete course",
        "operationId": "deleteCourse",
        "security": [{"BearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}],
        "responses": {
          "204": {"description": "Course deleted"},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      }
    },
...
    "/enrollments": {
      "get": {
        "tags": ["Enrollments"],
        "summary": "List enrollments",
        "operationId": "listEnrollments",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {"$ref": "#/components/parameters/Page"},
          {"$ref": "#/components/parameters/Limit"},
          {"name": "class_id", "in": "query", "schema": {"type": "string", "format": "uuid"}},
          {"name": "student_id", "in": "query", "schema": {"type": "string", "format": "uuid"}}
        ],
        "responses": {
          "200": {"description": "Enrollments retrieved", "content": {"application/json": {"schema": {"type": "object", "properties": {"data": {"type": "array", "items": {"$ref": "#/components/schemas/Enrollment"}}, "pagination": {"$ref": "#/components/schemas/PaginatedResult"}}}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"}
        }
      },
      "post": {
        "tags": ["Enrollments"],
        "summary": "Enroll student",
        "operationId": "enrollStudent",
        "security": [{"BearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/EnrollStudentRequest"}}}},
        "responses": {
          "201": {"description": "Student enrolled", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Enrollment"}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"}
        }
      }
    },
    "/enrollments/{id}": {
      "get": {
        "tags": ["Enrollments"],
        "summary": "Get enrollment by ID",
        "operationId": "getEnrollment",
        "security": [{"BearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}],
        "responses": {
          "200": {"description": "Enrollment retrieved", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Enrollment"}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      }
    },
    "/enrollments/{id}/transfer": {
      "post": {
        "tags": ["Enrollments"],
        "summary": "Transfer student to another class",
        "operationId": "transferStudent",
        "security": [{"BearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}, {"$ref": "#/components/parameters/IdempotencyKey"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "properties": {"target_class_id": {"type": "string", "format": "uuid"}}}}}},
        "responses": {
          "200": {"description": "Student transferred", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Enrollment"}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from /public/auth/login"
      }
    },
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "description": "Unique key to ensure idempotent requests. Prevents duplicate operations.",
        "required": false,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "Page": {
        "name": "page",
        "in": "query",
        "description": "Page number (1-indexed)",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "description": "Number of items per page",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        }
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Authentication required or token invalid/expired",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "UNAUTHORIZED",
                "message": "Authentication required",
                "details": {}
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "FORBIDDEN",
                "message": "Insufficient permissions to perform this action",
                "details": {}
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": {
                "code": "NOT_FOUND",
                "message": "Resource not found",
                "details": {}
              }
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "Class": {
        "type": "object",
        "properties": {
          "id": {"type": "string", "format": "uuid"},
          "tenant_id": {"type": "string", "format": "uuid"},
          "name": {"type": "string"},
          "grade_level": {"type": "string"},
          "section": {"type": "string"},
          "capacity": {"type": "integer"},
          "form_tutor_id": {"type": "string", "format": "uuid"},
          "status": {"type": "string", "enum": ["ACTIVE", "ARCHIVED"]},
          "student_count": {"type": "integer"},
          "created_at": {"type": "string", "format": "date-time"},
          "updated_at": {"type": "string", "format": "date-time"}
        }
      },
...
      "Course": {
        "type": "object",
        "properties": {
          "id": {"type": "string", "format": "uuid"},
          "tenant_id": {"type": "string", "format": "uuid"},
          "class_id": {"type": "string", "format": "uuid"},
          "name": {"type": "string"},
          "code": {"type": "string"},
          "description": {"type": "string"},
          "tutor_id": {"type": "string", "format": "uuid"},
          "tutor": {"$ref": "#/components/schemas/User"},
          "credits": {"type": "integer"},
          "periods_per_week": {"type": "integer"},
          "status": {"type": "string", "enum": ["ACTIVE", "ARCHIVED"]},
          "created_at": {"type": "string", "format": "date-time"},
          "updated_at": {"type": "string", "format": "date-time"}
        }
      },
      "CreateCourseRequest": {
        "type": "object",
        "required": ["class_id", "name", "code", "tutor_id"],
        "properties": {
          "class_id": {"type": "string", "format": "uuid"},
          "name": {"type": "string", "minLength": 1, "maxLength": 200},
          "code": {"type": "string", "minLength": 1, "maxLength": 20},
          "description": {"type": "string", "maxLength": 1000},
          "tutor_id": {"type": "string", "format": "uuid"},
          "credits": {"type": "integer", "minimum": 1},
          "periods_per_week": {"type": "integer", "minimum": 1, "maximum": 10}
        }
      },
      "UpdateCourseRequest": {
        "type": "object",
        "properties": {
          "name": {"type": "string", "minLength": 1, "maxLength": 200},
          "code": {"type": "string", "minLength": 1, "maxLength": 20},
          "description": {"type": "string", "maxLength": 1000},
          "tutor_id": {"type": "string", "format": "uuid"},
          "credits": {"type": "integer", "minimum": 1},
          "periods_per_week": {"type": "integer", "minimum": 1, "maximum": 10}
        }
      },
...
      "Enrollment": {
        "type": "object",
        "properties": {
          "id": {"type": "string", "format": "uuid"},
          "tenant_id": {"type": "string", "format": "uuid"},
          "student_id": {"type": "string", "format": "uuid"},
          "student": {"$ref": "#/components/schemas/User"},
          "class_id": {"type": "string", "format": "uuid"},
          "class": {"$ref": "#/components/schemas/Class"},
          "session_id": {"type": "string", "format": "uuid"},
          "status": {"type": "string", "enum": ["ACTIVE", "SUSPENDED", "WITHDRAWN", "GRADUATED", "TRANSFERRED"]},
          "enrolled_at": {"type": "string", "format": "date-time"},
          "withdrawn_at": {"type": "string", "format": "date-time"},
          "created_at": {"type": "string", "format": "date-time"},
          "updated_at": {"type": "string", "format": "date-time"}
        }
      },
      "EnrollStudentRequest": {
        "type": "object",
        "required": ["student_id", "class_id", "session_id"],
        "properties": {
          "student_id": {"type": "string", "format": "uuid"},
          "class_id": {"type": "string", "format": "uuid"},
          "session_id": {"type": "string", "format": "uuid"}
        }
      },
...
      "PaginatedResult": {
        "type": "object",
        "properties": {
          "total": {"type": "integer"},
          "page": {"type": "integer"},
          "limit": {"type": "integer"},
          "total_pages": {"type": "integer"},
          "has_next": {"type": "boolean"},
          "has_prev": {"type": "boolean"},
          "next_cursor": {"type": "string"},
          "prev_cursor": {"type": "string"}
        }
      },
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": {
                "type": "string",
                "description": "Machine-readable error code"
              },
              "message": {
                "type": "string",
                "description": "Human-readable error message"
              },
              "details": {
                "type": "object",
                "description": "Additional error context",
                "additionalProperties": true
              }
            }
          }
        }
      },
...
      "User": {
        "type": "object",
        "required": ["id", "role", "email", "first_name", "last_name", "status", "created_at", "updated_at"],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "NULL for SUPER_ADMIN"
          },
          "role": {
            "type": "string",
            "enum": ["SUPER_ADMIN", "ADMIN", "TUTOR", "STUDENT"]
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "first_name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "last_name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "middle_name": {
            "type": "string",
            "maxLength": 100,
            "nullable": true
          },
          "profile_photo": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": ["ACTIVE", "INACTIVE", "DEACTIVATED", "PENDING"]
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Array of resource:action permissions (e.g., 'tenant:create', 'user:delete')"
          },
          "notification_preferences": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotificationPreference"
            }
          },
          "last_login_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "deactivated_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "deactivation_reason": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "NotificationPreference": {
        "type": "object",
        "required": ["event_type", "in_app_enabled", "push_enabled", "email_enabled"],
        "properties": {
          "event_type": {
            "type": "string"
          },
          "in_app_enabled": {
            "type": "boolean"
          },
          "push_enabled": {
            "type": "boolean"
          },
          "email_enabled": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}